+++ CHANGELOG.md
@@ -8,7 +8,7 @@ This mod is distributed under Version 2 of the GPL.  A copy of the GPL is includ
 
 - reimplement a basic class filter for loot pages
 - (classic) move Lady Falther'ess from Razorfen Downs into extra list and mark as Scourge Invasions npc
-- (classic) add class icons for tokens from: Nax
+- (classic) add class icons for tokens from: Nax, AQ40
 - (bc) add bc support
 - (bc) add dungeons and raids
 - (bc) add Tier token items